+++ .gitignore
@@ -3,6 +3,7 @@ __azurite_db_*
 __blobstorage__/*
 __queuestorage__/*
 .DS_Store
+.venv/
 
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
