+++ DurableMultiAgentTemplate.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DurableMultiAgentTemplate", "DurableMultiAgentTemplate.csproj", "{B97B0F85-B39D-4768-806D-003066C3D491}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DurableMultiAgentTemplate", "DurableMultiAgentTemplate/DurableMultiAgentTemplate.csproj", "{B97B0F85-B39D-4768-806D-003066C3D491}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
