+++ DurableMultiAgentTemplate/Agent/AgentDefinition.cs
