+++ DurableMultiAgentTemplate/Agent/Orchestrator/SynthesizerPrompt.cs
