+++ DurableMultiAgentTemplate/Agent/Orchestrator/SynthesizerWithAdditionalInfoPrompt.cs
