+++ DurableMultiAgentTemplate/Json/JsonSchemaGenerator.cs
