+++ DurableMultiAgentTemplate/Model/AgentDeciderResult.cs
