+++ DurableMultiAgentTemplate/Model/AgentRequestDto.cs
