+++ DurableMultiAgentTemplate/Model/AgentResponseDto.cs
