+++ DurableMultiAgentTemplate/Model/AgentResponseWithAdditionalInfoDto.cs
