+++ DurableMultiAgentTemplate/Model/AgentResponseWithAdditionalInfoFormat.cs
