+++ DurableMultiAgentTemplate/Model/AppConfiguration.cs
