+++ DurableMultiAgentTemplate/Model/GetClimateRequest.cs
