+++ DurableMultiAgentTemplate/Model/GetDestinationSuggestRequest.cs
