+++ DurableMultiAgentTemplate/Model/GetHotelRequest.cs
