+++ DurableMultiAgentTemplate/Model/GetSightseeingSpotRequest.cs
