+++ DurableMultiAgentTemplate/Model/SourceGenerationContext.cs
