+++ DurableMultiAgentTemplate/Model/SubmitReservationRequest.cs
