+++ DurableMultiAgentTemplate/Model/SynthesizerRequest.cs
