+++ DurableMultiAgentTemplate/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "DurableMultiAgentTemplate": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7133",
+      "launchBrowser": false
+    }
+  }
+}
