--- Properties/launchSettings.json
@@ -1,9 +1,0 @@
-{
-  "profiles": {
-    "DurableMultiAgentTemplate": {
-      "commandName": "Project",
-      "commandLineArgs": "--port 7133",
-      "launchBrowser": false
-    }
-  }
-}
