--- local.settings.json
@@ -1,9 +1,0 @@
-{
-    "IsEncrypted": false,
-    "Values": {
-        "AzureWebJobsStorage": "UseDevelopmentStorage=true",
-        "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
-        "AppConfig:OpenAIEndpoint": "https://{your endpoint}.openai.azure.com/",
-        "AppConfig:OpenAIDeploy": "deploy"
-    }
-}
